+++ src/main/java/draylar/tiered/mixin/ItemStackMixin.java
@@ -5,6 +5,7 @@ import draylar.tiered.api.ModifierUtils;
 import draylar.tiered.api.PotentialAttribute;
 import net.minecraft.component.type.AttributeModifierSlot;
 import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.item.ItemStack;
@@ -39,6 +40,7 @@ public abstract class ItemStackMixin {
     }
 
 
+    // Used for tooltip
     @Inject(method = "Lnet/minecraft/item/ItemStack;applyAttributeModifier(Lnet/minecraft/component/type/AttributeModifierSlot;Ljava/util/function/BiConsumer;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/enchantment/EnchantmentHelper;applyAttributeModifiers(Lnet/minecraft/item/ItemStack;Lnet/minecraft/component/type/AttributeModifierSlot;Ljava/util/function/BiConsumer;)V"))
     private void applyAttributeModifierMixin(AttributeModifierSlot slot, BiConsumer<RegistryEntry<EntityAttribute>, EntityAttributeModifier> attributeModifierConsumer, CallbackInfo info) {
         applyAttributeModifier(null, slot, attributeModifierConsumer);
@@ -76,8 +78,7 @@ public abstract class ItemStackMixin {
                         // optional equipment slots are valid ONLY IF the equipment slot is valid for the thing
                         if (equipmentSlot != null && optionalEquipmentSlots.contains(equipmentSlot) && Tiered.isPreferredEquipmentSlot(itemStack, equipmentSlot)) {
                             template.applyModifiers(equipmentSlot, attributeModifierConsumer);
-                        } else if (attributeModifierSlot != null) {
-
+                        } else if (attributeModifierSlot != null && attributeModifierSlot != AttributeModifierSlot.ANY && attributeModifierSlot != AttributeModifierSlot.HAND) {
                             Optional<EquipmentSlot> optional = Arrays.stream(template.getOptionalEquipmentSlots()).filter(attributeModifierSlot::matches).findFirst();
                             if (optional.isPresent() && Tiered.isPreferredEquipmentSlot(itemStack, optional.get())) {
                                 template.applyModifiers(optional.get(), attributeModifierConsumer);