+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Fixed shield tier
+- Fixed mainhand tier tooltip
 ### Changed:
 -